# places_app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doc API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Список пользователей",
                "operationId": "6f8f0134e30ae7557795c2a8f042fd83",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "14"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "user"
                                                    },
                                                    "login": {
                                                        "type": "string",
                                                        "example": "andrew"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Создание пользователя",
                "operationId": "6da8a20df7b3e41929be6abac6e75525",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "login": {
                                                "type": "string",
                                                "example": "andrew"
                                            },
                                            "role": {
                                                "type": "string",
                                                "example": "user"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password1234"
                                            },
                                            "password_confirmation": {
                                                "type": "string",
                                                "example": "password1234"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "user"
                                                },
                                                "login": {
                                                    "type": "string",
                                                    "example": "andrew"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Запись о пользователе",
                "operationId": "06188816e296c47e37dca39a0d2b792f",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "example": 14
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "user"
                                                },
                                                "login": {
                                                    "type": "string",
                                                    "example": "andrew"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}/favorites": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Список желаемого пользователя",
                "operationId": "683a98a5e6b2a3013ab57c95f70ca856",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "example": 14
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "14"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Some place"
                                                    },
                                                    "longitude": {
                                                        "type": "number",
                                                        "example": "14.324987"
                                                    },
                                                    "latitude": {
                                                        "type": "number",
                                                        "example": "-60.324987"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}/place/{place}/add": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Добавление места в список желаемых заданного пользователя",
                "operationId": "819037f22539187abd7a999619d37514",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "example": 14
                    },
                    {
                        "name": "place",
                        "in": "path",
                        "description": "ID места",
                        "required": true,
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some place"
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "example": "14.324987"
                                                },
                                                "latitude": {
                                                    "type": "number",
                                                    "example": "-60.324987"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "No more items can be added | This item has already been added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "422"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No more items can be added"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}/place/{place}/remove": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Удаление места из список желаемого заданного пользователя",
                "operationId": "832450d7bfdfe70a0f8dce0793cdbe9c",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "example": 14
                    },
                    {
                        "name": "place",
                        "in": "path",
                        "description": "ID места",
                        "required": true,
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some place"
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "example": "14.324987"
                                                },
                                                "latitude": {
                                                    "type": "number",
                                                    "example": "-60.324987"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "This item is not included in the list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "422"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "This item is not included in the list"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}